refactor: top-level stereo module, gate and consolidate eval
- promote stereo correspondence from estimation/ to kornia_slam::stereo
  (it produces per-keypoint depth, not pose Estimates)
- gate orb_slam EuRoC evaluation behind --evaluate; add --eval-out to
  pick the CSV output directory (default: cwd, created if missing)
- collapse the orb_slam evaluation/ dir into a single evaluation.rs and
  lift the ~50-line report block out of main.rs
- note the Bouguet rectifier's eventual home (kornia-imgproc upstream)
