+++ crates/kornia-slam/src/estimation/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod map_projection;
 pub mod pnp;
-pub mod stereo;
 pub mod two_view;
 
 use kornia_3d::pose::Pose3d;
