--- examples/orb_slam/src/evaluation/mod.rs
@@ -1,8 +1,0 @@
-pub mod alignment;
-pub mod association;
-pub mod metrics;
-pub mod types;
-
-pub use alignment::align_sim3;
-pub use association::associate_gt;
-pub use metrics::{compute_ate, compute_drift, compute_rpe};
