+++ examples/orb_slam/src/main.rs
@@ -33,13 +33,13 @@ mod tui;
 mod utils;
 use crate::datasets::euroc::GroundTruthPose;
 use config::PipelineConfig;
-use evaluation::{align_sim3, associate_gt, compute_ate, compute_drift, compute_rpe};
+use evaluation::associate_gt;
 use kornia_3d::pose::Pose3d;
 use kornia_algebra::Vec3F64;
 use kornia_image::{Image, ImageSize, InterpolationMode};
 use kornia_imgproc::resize::resize_fast_mono;
 use kornia_slam::Frame;
-use kornia_slam::estimation::stereo::{StereoMatchConfig, compute_stereo_matches};
+use kornia_slam::stereo::{StereoMatchConfig, compute_stereo_matches};
 use kornia_tensor::CpuAllocator;
 use pipeline::Pipeline;
 #[cfg(feature = "oakd")]
@@ -104,6 +104,15 @@ struct EurocCmd {
     /// rectify cam0+cam1 and compute per-keypoint stereo depth
     #[argh(switch)]
     stereo: bool,
+
+    /// after the run, align the trajectory to ground truth and report
+    /// ATE/RPE/drift (writes kornia_slam_raw.csv and kornia_slam_aligned.csv)
+    #[argh(switch)]
+    evaluate: bool,
+
+    /// directory for the evaluation CSVs (created if missing; default: current dir)
+    #[argh(option, default = "String::from(\".\")")]
+    eval_out: String,
 }
 
 /// Run on a bubbaloop MCAP recording.
@@ -245,9 +254,13 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let tui_active = !args.no_tui;
 
     // ── Source ─────────────────────────────────────────────────────────────
+    let mut evaluate = false;
+    let mut eval_out = String::from(".");
     let (mut source, euroc_gt): (Box<dyn FrameSource>, Option<Vec<GroundTruthPose>>) =
         match args.source {
             SourceCmd::Euroc(e) => {
+                evaluate = e.evaluate;
+                eval_out = e.eval_out.clone();
                 let src = if e.stereo {
                     EurocSource::open_stereo(&e.data, e.start_frame, e.max_frames)?
                 } else {
@@ -365,7 +378,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     } else {
         None
     };
-    let euroc_samples_meta = euroc_gt.is_some();
     let (mut est_positions, mut gt_positions): (Vec<Vec3F64>, Vec<Vec3F64>) =
         (Vec::new(), Vec::new());
     // ── Main loop ──────────────────────────────────────────────────────────
@@ -460,8 +472,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         let traj_pt = trajectory_point_from_pose(&result.pose_world_to_cam);
         trajectory.push(traj_pt);
 
-        // Evaluation collection (EuRoC only).
-        if euroc_samples_meta {
+        // Evaluation collection (EuRoC, --evaluate only).
+        if evaluate {
             let est_pos = Vec3F64::new(traj_pt[0] as f64, traj_pt[1] as f64, traj_pt[2] as f64);
             est_positions.push(est_pos);
 
@@ -535,56 +547,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     eprintln!(
         "Done. Final map: total={total_pts}  active={active_pts}  obs_per_active_mp={obs_mean:.2}  max_obs={obs_max}"
     );
-    // ── Trajectory evaluation (EuRoC only) ─────────────────────────────────
-    if euroc_samples_meta && est_positions.len() >= 2 && gt_positions.len() == est_positions.len() {
-        // Save raw (unaligned, monocular scale) trajectory.
-        {
-            use std::io::Write;
-            let mut raw_file = std::fs::File::create("kornia_slam_raw.csv")?;
-            writeln!(raw_file, "timestamp_sec,x,y,z")?;
-            // We no longer have sample timestamps here; reconstruct index from source.
-            // Write position index as a proxy — replace with real timestamps if
-            // EurocSource exposes them via a `timestamps()` method.
-            for (i, pos) in est_positions.iter().enumerate() {
-                writeln!(raw_file, "{i},{:.9},{:.9},{:.9}", pos.x, pos.y, pos.z)?;
-            }
-            eprintln!("Saved kornia_slam_raw.csv");
-        }
-
-        let alignment = align_sim3(&est_positions, &gt_positions);
-
-        let aligned_est: Vec<Vec3F64> = est_positions
-            .iter()
-            .map(|p| alignment.translation + (alignment.rotation * *p) * alignment.scale)
-            .collect();
-
-        // Save aligned trajectory.
-        {
-            use std::io::Write;
-            let mut aligned_file = std::fs::File::create("kornia_slam_aligned.csv")?;
-            writeln!(aligned_file, "idx,x,y,z,gt_x,gt_y,gt_z")?;
-            for (i, (pos, gt)) in aligned_est.iter().zip(gt_positions.iter()).enumerate() {
-                writeln!(
-                    aligned_file,
-                    "{i},{:.9},{:.9},{:.9},{:.9},{:.9},{:.9}",
-                    pos.x, pos.y, pos.z, gt.x, gt.y, gt.z
-                )?;
-            }
-            eprintln!("Saved kornia_slam_aligned.csv");
-        }
-
-        let ate = compute_ate(&aligned_est, &gt_positions);
-        let rpe = compute_rpe(&aligned_est, &gt_positions);
-        let drift = compute_drift(&aligned_est, &gt_positions);
-
-        eprintln!();
-        eprintln!("================ Trajectory Evaluation ================");
-        eprintln!("Frames evaluated : {}", aligned_est.len());
-        eprintln!("Scale            : {:.6}", alignment.scale);
-        eprintln!("ATE RMSE         : {:.4} m", ate);
-        eprintln!("RPE RMSE         : {:.4} m", rpe);
-        eprintln!("Final Drift      : {:.4} %", drift * 100.0);
-        eprintln!("=======================================================");
+    // ── Trajectory evaluation (EuRoC, --evaluate only) ─────────────────────
+    if evaluate {
+        evaluation::report(&est_positions, &gt_positions, std::path::Path::new(&eval_out))?;
     }
     Ok(())
 }
